requirements.txt
langchain
wikipedia
langchain-openai
langchain-community
langchain-anthropic
python-dotenv
pydantic